mi0.3.0 - Connection to database
- `app.py` now can infer the db by its own.
- `.gitignore` updated for pycache and journal files.

This minor update brings full coverage of db connection, which means
that the app first looks for a production database and, if fails,
try to create a new one for testing purposes.

This is the final update prior to start working at the docker update
and the change of the metal infrastructure.