mi0.4.0 - Added docker file for server deployment
- `Dockerfile`: added build instructions for arm and amd
architectures, both in development mode.
- `app.py`: updated production port for integrating with PaaS
products.

with this change, the app integrates well with one out of two
platforms in sight. Next update must look forward to deliver an app
in a PaaS product which must be compatible with volumes. It will be
needed to add support for the database to be downloaded otherwise.